+++ client/onboarding-prototype/translations/descriptions.tsx
